Introduced BaseAlarmProcessor
