+++ application/src/main/java/org/thingsboard/server/service/edge/rpc/processor/alarm/AlarmEdgeProcessor.java
@@ -22,25 +22,19 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.thingsboard.common.util.JacksonUtil;
 import org.thingsboard.server.common.data.EdgeUtils;
-import org.thingsboard.server.common.data.EntityType;
 import org.thingsboard.server.common.data.alarm.Alarm;
-import org.thingsboard.server.common.data.alarm.AlarmSeverity;
-import org.thingsboard.server.common.data.alarm.AlarmStatus;
 import org.thingsboard.server.common.data.edge.EdgeEvent;
 import org.thingsboard.server.common.data.edge.EdgeEventActionType;
 import org.thingsboard.server.common.data.edge.EdgeEventType;
 import org.thingsboard.server.common.data.id.AlarmId;
 import org.thingsboard.server.common.data.id.EdgeId;
-import org.thingsboard.server.common.data.id.EntityId;
 import org.thingsboard.server.common.data.id.TenantId;
 import org.thingsboard.server.common.data.page.PageData;
 import org.thingsboard.server.common.data.page.PageLink;
 import org.thingsboard.server.gen.edge.v1.AlarmUpdateMsg;
 import org.thingsboard.server.gen.edge.v1.DownlinkMsg;
-import org.thingsboard.server.gen.edge.v1.UpdateMsgType;
 import org.thingsboard.server.gen.transport.TransportProtos;
 import org.thingsboard.server.queue.util.TbCoreComponent;
-import org.thingsboard.server.service.edge.rpc.processor.BaseEdgeProcessor;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -49,106 +43,18 @@ import java.util.UUID;
 @Component
 @Slf4j
 @TbCoreComponent
-public class AlarmEdgeProcessor extends BaseEdgeProcessor {
-
-    public ListenableFuture<Void> processAlarmFromEdge(TenantId tenantId, AlarmUpdateMsg alarmUpdateMsg) {
-        log.trace("[{}] processAlarmFromEdge [{}]", tenantId, alarmUpdateMsg);
-        EntityId originatorId = getAlarmOriginator(tenantId, alarmUpdateMsg.getOriginatorName(),
-                EntityType.valueOf(alarmUpdateMsg.getOriginatorType()));
-        if (originatorId == null) {
-            log.warn("Originator not found for the alarm msg {}", alarmUpdateMsg);
-            return Futures.immediateFuture(null);
-        }
-        try {
-            Alarm existentAlarm = alarmService.findLatestByOriginatorAndType(tenantId, originatorId, alarmUpdateMsg.getType()).get();
-            switch (alarmUpdateMsg.getMsgType()) {
-                case ENTITY_CREATED_RPC_MESSAGE:
-                case ENTITY_UPDATED_RPC_MESSAGE:
-                    if (existentAlarm == null || existentAlarm.getStatus().isCleared()) {
-                        existentAlarm = new Alarm();
-                        existentAlarm.setTenantId(tenantId);
-                        existentAlarm.setType(alarmUpdateMsg.getName());
-                        existentAlarm.setOriginator(originatorId);
-                        existentAlarm.setSeverity(AlarmSeverity.valueOf(alarmUpdateMsg.getSeverity()));
-                        existentAlarm.setStartTs(alarmUpdateMsg.getStartTs());
-                        existentAlarm.setClearTs(alarmUpdateMsg.getClearTs());
-                        existentAlarm.setPropagate(alarmUpdateMsg.getPropagate());
-                    }
-                    existentAlarm.setStatus(AlarmStatus.valueOf(alarmUpdateMsg.getStatus()));
-                    existentAlarm.setAckTs(alarmUpdateMsg.getAckTs());
-                    existentAlarm.setEndTs(alarmUpdateMsg.getEndTs());
-                    existentAlarm.setDetails(JacksonUtil.OBJECT_MAPPER.readTree(alarmUpdateMsg.getDetails()));
-                    alarmService.createOrUpdateAlarm(existentAlarm);
-                    break;
-                case ALARM_ACK_RPC_MESSAGE:
-                    if (existentAlarm != null) {
-                        alarmService.ackAlarm(tenantId, existentAlarm.getId(), alarmUpdateMsg.getAckTs());
-                    }
-                    break;
-                case ALARM_CLEAR_RPC_MESSAGE:
-                    if (existentAlarm != null) {
-                        alarmService.clearAlarm(tenantId, existentAlarm.getId(),
-                                JacksonUtil.OBJECT_MAPPER.readTree(alarmUpdateMsg.getDetails()), alarmUpdateMsg.getAckTs());
-                    }
-                    break;
-                case ENTITY_DELETED_RPC_MESSAGE:
-                    if (existentAlarm != null) {
-                        alarmService.deleteAlarm(tenantId, existentAlarm.getId());
-                    }
-                    break;
-            }
-            return Futures.immediateFuture(null);
-        } catch (Exception e) {
-            log.error("Failed to process alarm update msg [{}]", alarmUpdateMsg, e);
-            return Futures.immediateFailedFuture(new RuntimeException("Failed to process alarm update msg", e));
-        }
-    }
-
-    private EntityId getAlarmOriginator(TenantId tenantId, String entityName, EntityType entityType) {
-        switch (entityType) {
-            case DEVICE:
-                return deviceService.findDeviceByTenantIdAndName(tenantId, entityName).getId();
-            case ASSET:
-                return assetService.findAssetByTenantIdAndName(tenantId, entityName).getId();
-            case ENTITY_VIEW:
-                return entityViewService.findEntityViewByTenantIdAndName(tenantId, entityName).getId();
-            default:
-                return null;
-        }
-    }
+public class AlarmEdgeProcessor extends BaseAlarmProcessor {
 
     public DownlinkMsg convertAlarmEventToDownlink(EdgeEvent edgeEvent) {
-        AlarmId alarmId = new AlarmId(edgeEvent.getEntityId());
-        DownlinkMsg downlinkMsg = null;
-        UpdateMsgType msgType = getUpdateMsgType(edgeEvent.getAction());
-        switch (edgeEvent.getAction()) {
-            case ADDED:
-            case UPDATED:
-            case ALARM_ACK:
-            case ALARM_CLEAR:
-                try {
-                    Alarm alarm = alarmService.findAlarmByIdAsync(edgeEvent.getTenantId(), alarmId).get();
-                    if (alarm != null) {
-                        downlinkMsg = DownlinkMsg.newBuilder()
-                                .setDownlinkMsgId(EdgeUtils.nextPositiveInt())
-                                .addAlarmUpdateMsg(alarmMsgConstructor.constructAlarmUpdatedMsg(edgeEvent.getTenantId(), msgType, alarm))
-                                .build();
-                    }
-                } catch (Exception e) {
-                    log.error("Can't process alarm msg [{}] [{}]", edgeEvent, msgType, e);
-                }
-                break;
-            case DELETED:
-                Alarm alarm = JacksonUtil.OBJECT_MAPPER.convertValue(edgeEvent.getBody(), Alarm.class);
-                AlarmUpdateMsg alarmUpdateMsg =
-                        alarmMsgConstructor.constructAlarmUpdatedMsg(edgeEvent.getTenantId(), msgType, alarm);
-                downlinkMsg = DownlinkMsg.newBuilder()
-                        .setDownlinkMsgId(EdgeUtils.nextPositiveInt())
-                        .addAlarmUpdateMsg(alarmUpdateMsg)
-                        .build();
-                break;
+        AlarmUpdateMsg alarmUpdateMsg =
+                convertAlarmEventToAlarmMsg(edgeEvent.getTenantId(), edgeEvent.getEntityId(), edgeEvent.getAction(), edgeEvent.getBody());
+        if (alarmUpdateMsg != null) {
+            return DownlinkMsg.newBuilder()
+                    .setDownlinkMsgId(EdgeUtils.nextPositiveInt())
+                    .addAlarmUpdateMsg(alarmUpdateMsg)
+                    .build();
         }
-        return downlinkMsg;
+        return null;
     }
 
     public ListenableFuture<Void> processAlarmNotification(TenantId tenantId, TransportProtos.EdgeNotificationMsgProto edgeNotificationMsg) throws JsonProcessingException {
